Create Online Exam System Managment
"This Git repository contains the source code and documentation for the online exam management system developed as part of the Digirise Infolabs internship project. The web-based platform enables administrators to create and grade customizable exams, while students can take the exams and view their results. The repository includes all the files necessary for building and running the system. To contribute, create a new branch from 'develop' and submit a pull request for review."